+++ examples/CMakeLists.txt
@@ -1,7 +1,7 @@
 INCLUDE_DIRECTORIES(${LIBGIT2_INCLUDES})
 INCLUDE_DIRECTORIES(SYSTEM ${LIBGIT2_SYSTEM_INCLUDES})
 
-FILE(GLOB_RECURSE SRC_EXAMPLE_GIT2 network/*.c network/*.h common.?)
+FILE(GLOB_RECURSE SRC_EXAMPLE_GIT2 network/*.c common.?)
 ADD_EXECUTABLE(cgit2 ${SRC_EXAMPLE_GIT2})
 SET_TARGET_PROPERTIES(cgit2 PROPERTIES C_STANDARD 90)
