+++ examples/network/clone.c
@@ -1,13 +1,4 @@
-#include "common.h"
-#include <git2.h>
-#include <git2/clone.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#ifndef _WIN32
-# include <pthread.h>
-# include <unistd.h>
-#endif
+#include "../common.h"
 
 typedef struct progress_data {
 	git_transfer_progress fetch_progress;
