+++ examples/network/ls-remote.c
@@ -1,8 +1,4 @@
-#include <git2.h>
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include "common.h"
+#include "../common.h"
 
 static int use_remote(git_repository *repo, char *name)
 {
